+++ src/DrupalTest/BrowserTestCase.php
@@ -305,7 +305,7 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
   /**
    * Assert the status code at a given URL equals expected.
    *
-   * @param int $expected_status_code
+   * @param int $status_code
    *   The expected status code.
    * @param string $url
    *   The URL to access.
@@ -313,18 +313,18 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
    * @return \AKlump\DrupalTest\BrowserTestCase
    *   Self for chaining.
    */
-  public function assertUrlStatusCodeEquals($expected_status_code, $url) {
-    $assertion = function ($response) use ($expected_status_code) {
-      $this->assertEquals($expected_status_code, $response->getStatusCode(), "HTTP status code of " . $response->getStatusCode() . " does not equal expected $expected_status_code.");
+  public function assertUrlStatusCodeEquals($status_code, $url) {
+    $assertion = function ($response) use ($status_code) {
+      static::assertThat($status_code == $response->getStatusCode(), static::isTrue(), "HTTP status code of " . $response->getStatusCode() . " does not equal expected $status_code.");
     };
 
-    return $this->requestThenAssert($assertion, $url, 'HEAD', $assertion);
+    return $this->requestThenAssert($assertion, $url, [], 'HEAD', $assertion);
   }
 
   /**
    * Assert the status code at a given URL does not equal a value.
    *
-   * @param int $expected_status_code
+   * @param int $status_code
    *   The status code to not equal.
    * @param string $url
    *   The URL to access.
@@ -332,18 +332,18 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
    * @return \AKlump\DrupalTest\BrowserTestCase
    *   Self for chaining.
    */
-  public function assertUrlStatusCodeNotEquals($expected_status_code, $url) {
-    $assertion = function ($response) use ($expected_status_code) {
-      $this->assertNotEquals($expected_status_code, $response->getStatusCode(), "HTTP status code of " . $response->getStatusCode() . " should not equal $expected_status_code.");
+  public function assertUrlStatusCodeNotEquals($status_code, $url) {
+    $assertion = function ($response) use ($status_code) {
+      static::assertThat($status_code != $response->getStatusCode(), static::isTrue(), "HTTP status code of " . $response->getStatusCode() . " should not equal $status_code.");
     };
 
-    return $this->requestThenAssert($assertion, $url, 'HEAD', $assertion);
+    return $this->requestThenAssert($assertion, $url, [], 'HEAD', $assertion);
   }
 
   /**
    * Assert an URL returns a certain content type header.
    *
-   * @param string $expected_content_type
+   * @param string $content_type
    *   The expected mime-type in the content-type header.
    * @param string $url
    *   The URL to access.
@@ -351,12 +351,33 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
    * @return \AKlump\DrupalTest\BrowserTestCase
    *   Self for chaining.
    */
-  public function assertUrlContentTypeEquals($expected_content_type, $url) {
-    return $this->requestThenAssert(function ($response) use ($expected_content_type) {
+  public function assertUrlContentTypeEquals($content_type, $url) {
+    return $this->requestThenAssert(function ($response) use ($content_type) {
       $actual = $response->getHeader('content-type');
       $actual = array_pop($actual);
-      $this->assertEquals($expected_content_type, $actual, "Actual content type \"$actual\" does not match expected \"$expected_content_type\".");
+      static::assertThat($content_type == $actual, static::isTrue(), "Actual content type \"$actual\" does not match expected \"$content_type\".");
     }, $url);
+  }
+
+  /**
+   * Assert that one URL redirects to another.
+   *
+   * @param string $final_url
+   *   The expected destinaton URL.
+   * @param string $url
+   *   The URL that will redirect.
+   *
+   * @return $this
+   *   Self for chaining.
+   */
+  public function assertUrlRedirectsTo($final_url, $url) {
+    return $this->requestThenAssert(function ($response) use ($final_url, $url) {
+      if ($location = $response->getHeader('location')) {
+        $location = array_pop($location);
+        $final_url = $this->resolveUrl($final_url, TRUE);
+      }
+      static::assertThat($final_url === $location, static::isTrue(), "Failed asserting that $url redirects to $final_url");
+    }, $url, ['allow_redirects' => FALSE]);
   }
 
   /**
@@ -368,14 +389,18 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
    *   Callback that receives ($request) and must make one or more assertions.
    * @param string $url
    *   The URL to make a head request against.
+   * @param array $client_options
+   *   An optional array to send to getClient().
    * @param string $method
    *   The http method to use, defaults to 'HEAD'.
+   * @param callable|null $on_fail
+   *   Optional, custom handler if the request throws BadResponseException.
    *
    * @return $this
    *   Self for chaining.
    */
-  protected function requestThenAssert(callable $callback, $url, $method = 'HEAD', callable $on_fail = NULL) {
-    $client = $this->getClient();
+  protected function requestThenAssert(callable $callback, $url, array $client_options = [], $method = 'HEAD', callable $on_fail = NULL) {
+    $client = $this->getClient($client_options);
     $url = $this->resolveUrl($url);
     if (empty($url)) {
       throw new \RuntimeException("\$url cannot be empty");
@@ -495,7 +520,7 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
   /**
    * Empty the cookie jar to create a new browsing session.
    */
-  public static function emptyCookieJar() {
+  public function emptyCookieJar() {
     static::$cookieJar = new CookieJar();
   }
 
@@ -505,7 +530,7 @@ abstract class BrowserTestCase extends ParentBrowserTestCase {
    * @return \GuzzleHttp\Cookie\CookieJar
    *   The current cookie jar instance.
    */
-  public static function getCookieJar() {
+  public function getCookieJar() {
     if (!static::$cookieJar) {
       static::emptyCookieJar();
     }