+++ documentation/source/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.6.0
  
+* `getDrupalCommandsClient` is no longer static.
+* `getXmlClient` is no longer static.
+* `getHtmlClient` is no longer static.
+* `getJsonClient` is no longer static.
+* `emptyCookieJar` is no longer static. 
 * `assertHttpStatusCodeAtUrl` renamed to `assertUrlStatusCodeEquals`.
 * `assertContentType` has been replaced with `assertUrlContentTypeEquals`.